+++ source/Public/Get-RelativeNodeFileName.ps1
@@ -1,0 +1,13 @@
+function Get-RelativeNodeFileName
+{
+    param (
+        [Parameter(Mandatory = $true)]
+        [System.IO.FileInfo]$File
+    )
+    
+    $p = Resolve-Path -Path $File.FullName -Relative
+    $p = $p -split '\\'
+    $p[-1] = [System.IO.Path]::GetFileNameWithoutExtension($p[-1])
+    $p[2..($p.Length - 1)] -join '\'
+
+}