+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Configurable Header and Footer
 - Content is now evaluated with RegEx + PowerShell Parser
 - Gives access to Node and Datum variable
+- Added function 'Get-RelativeNodeFileName'
 
 ## [0.1.1] - 2020-08-25